+++ libsrc/pylith/bc/NeumannUserFn.cc
@@ -173,25 +173,6 @@ pylith::bc::NeumannUserFn::_getAuxiliaryFactory(void) {
 
 
 // ------------------------------------------------------------------------------------------------
-// Update kernel constants.
-void
-pylith::bc::NeumannUserFn::_updateKernelConstants(const PylithReal dt) {
-    PYLITH_METHOD_BEGIN;
-    PYLITH_COMPONENT_DEBUG("_updateKernelConstants(dt="<<dt<<")");
-
-    if (6 != _kernelConstants.size()) { _kernelConstants.resize(6);}
-    _kernelConstants[0] = _refDir1[0];
-    _kernelConstants[1] = _refDir1[1];
-    _kernelConstants[2] = _refDir1[2];
-    _kernelConstants[3] = _refDir2[0];
-    _kernelConstants[4] = _refDir2[1];
-    _kernelConstants[5] = _refDir2[2];
-
-    PYLITH_METHOD_END;
-} // _updateKernelConstants
-
-
-// ------------------------------------------------------------------------------------------------
 // Set kernels for residual.
 void
 pylith::bc::_NeumannUserFn::setKernelsResidual(pylith::feassemble::IntegratorBoundary* integrator,